tools/file_processors.py
# tools/file_processors.py

import fitz  # PyMuPDF
from langchain.tools import tool

@tool
def pdf_reader_tool(file_path: str) -> str:
    """
    Reads a PDF file and returns its text content.
    The agent should provide the full, correct path to the PDF file.
    """
    try:
        doc = fitz.open(file_path)
        text = ""
        for page in doc:
            text += page.get_text()
        doc.close()
        return text
    except Exception as e:
        return f"Error reading PDF file at {file_path}: {e}. Please ensure the path is correct and the file is a valid PDF."

@tool
def meter_reader_tool(file_path: str) -> str:
    """
    Reads a meter value from an image file.
    The agent should provide the full, correct path to the image file (JPG, PNG).

    !! IMPORTANT HARD TRUTH !!
    This is a placeholder. Real-world OCR is complex. To make this work reliably,
    you would need to replace the logic here with a call to a powerful vision model
    like Gemini 1.5 Pro's native vision capabilities or a dedicated OCR API.
    The agent's LLM is a language model, not a vision model, so it cannot "see" the image itself.
    This tool's job is to be its eyes.
    """
    #
    # --- START OF PLACEHOLDER ---
    # In a real implementation, you would do this:
    # 1. Load the image using a library like Pillow.
    # 2. Encode the image (e.g., to base64).
    # 3. Send it to a multimodal model API (like the Gemini API with an image input).
    # 4. Parse the numeric reading from the model's response.
    #
    # For now, this tool will return a fixed value and a warning.
    #
    print("\n⚠️ WARNING: Using placeholder OCR tool. This is not reading the image file.")
    try:
        # We can try to get the user to input it manually for this demo version
        reading = input(f"  > Please manually enter the meter reading from the image at '{file_path}': ")
        return f"Manual reading provided: {float(reading)}"
    except ValueError:
        return "Error: Invalid number entered for manual meter reading."
    except Exception as e:
        return f"Error processing image file at {file_path}: {e}"
    # --- END OF PLACEHOLDER ---
